# Backend/routes/openapi.yaml
openapi: 3.0.3
info:
  title: Sabhyata Foundation API
  description: Comprehensive API documentation for the Sabhyata Foundation backend, handling authentication, admin operations, public endpoints, and payments.
  version: 1.0.0
  contact:
    name: Sabhyata Foundation
    email: support@sabhyata-foundation.org
servers:
  - url: http://localhost:5000/api
    description: Local development server
  - url: https://sabhyata-foundation-api.example.com/api
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
        error:
          type: string
      required:
        - success
...
    SeatData:
      type: object
      properties:
        seatId:
          type: string
        row:
          type: string
        number:
          type: number
        section:
          type: string
        category:
          type: string
          enum: [VIP, Premium, Gold, Silver, Bronze]
        price:
          type: number
          minimum: 0
        status:
          type: string
          enum: [available, booked, locked, blocked]
        lockedBy:
          type: string
        lockedAt:
          type: string
          format: date-time
        coords:
          type: object
          properties:
            x:
              type: number
            y:
              type: number
          required:
            - x
            - y
      required:
        - seatId
        - row
        - number
        - section
        - category
        - price
        - coords
    SeatLayout:
      type: object
      properties:
        _id:
          type: string
        event_id:
          type: string
        layout_data:
          type: array
          items:
            $ref: '#/components/schemas/SeatData'
        total_seats:
          type: number
        available_seats:
          type: number
        booked_seats:
          type: number
        layout_name:
          type: string
        created_by:
          type: string
        is_published:
          type: boolean
        stage:
          type: object
          properties:
            x:
              type: number
            y:
              type: number
            width:
              type: number
            height:
              type: number
            label:
              type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Ticket:
      type: object
      properties:
        ticketId:
          type: string
        type:
          type: string
          enum: [adult, child]
        price:
          type: number
        isUsed:
          type: boolean
      required:
        - type
        - price
    Booking:
      type: object
      properties:
        _id:
          type: string
        bookingReference:
          type: string
        event:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
        language:
          type: string
        seats:
          type: array
          items:
            $ref: '#/components/schemas/SeatData'
        tickets:
          type: array
          items:
            $ref: '#/components/schemas/Ticket'
        totalAmount:
          type: number
        contactInfo:
          type: object
          properties:
            name:
              type: string
            email:
              type: string
            phone:
              type: string
        paymentMethod:
          type: string
          enum: [card, upi, netbanking, wallet]
        notes:
          type: string
        status:
          type: string
          enum: [pending, confirmed, cancelled]
        paymentStatus:
          type: string
          enum: [pending, paid, failed]
        sessionId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - event
        - date
        - time
        - tickets
        - totalAmount
...
paths:
...
  /admin/seat-layouts:
    post:
      summary: Create a seat layout
      tags:
        - Seat Layouts
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                eventId:
                  type: string
                seats:
                  type: array
                  items:
                    type: object
                    properties:
                      seatId:
                        type: string
                      status:
                        type: string
                        enum: [available, held, booked, locked]
              required:
                - eventId
                - seats
      responses:
        '201':
          description: Seat layout created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatLayout'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (not admin)
  /admin/seat-layouts/{event_id}:
    get:
      summary: Get seat layout for an event
      tags:
        - Seat Layouts
      security:
        - bearerAuth: []
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Seat layout details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatLayout'
        '404':
          description: Seat layout not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (not admin)
    put:
      summary: Update a seat layout
      tags:
        - Seat Layouts
      security:
        - bearerAuth: []
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                seats:
                  type: array
                  items:
                    type: object
                    properties:
                      seatId:
                        type: string
                      status:
                        type: string
                        enum: [available, held, booked, locked]
      responses:
        '200':
          description: Seat layout updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatLayout'
        '400':
          description: Invalid input
        '404':
          description: Seat layout not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (not admin)
    delete:
      summary: Delete a seat layout
      tags:
        - Seat Layouts
      security:
        - bearerAuth: []
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Seat layout deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Seat layout not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (not admin)
  /admin/seat-layouts/{event_id}/hold-seats:
    post:
      summary: Hold seats for an event
      tags:
        - Seat Layouts
      security:
        - bearerAuth: []
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                seatIds:
                  type: array
                  items:
                    type: string
              required:
                - seatIds
      responses:
        '200':
          description: Seats held
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatLayout'
        '400':
          description: Invalid input
        '404':
          description: Seat layout not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (not admin)
  /admin/seat-layouts/{event_id}/release-seats:
    post:
      summary: Release held seats
      tags:
        - Seat Layouts
      security:
        - bearerAuth: []
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                seatIds:
                  type: array
                  items:
                    type: string
              required:
                - seatIds
      responses:
        '200':
          description: Seats released
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatLayout'
        '400':
          description: Invalid input
        '404':
          description: Seat layout not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (not admin)
  /admin/seat-layouts/{event_id}/publish:
    post:
      summary: Publish a seat layout
      tags:
        - Seat Layouts
      security:
        - bearerAuth: []
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Seat layout published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatLayout'
        '404':
          description: Seat layout not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (not admin)
  /admin/seat-layouts/{event_id}/availability:
    get:
      summary: Get seat availability
      tags:
        - Seat Layouts
      security:
        - bearerAuth: []
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Seat availability details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatLayout'
        '404':
          description: Seat layout not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (not admin)
  /admin/seat-layouts/{id}/book:
    post:
      summary: Book seats for an event
      tags:
        - Seat Layouts
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                seatIds:
                  type: array
                  items:
                    type: string
              required:
                - seatIds
      responses:
        '200':
          description: Seats booked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: Invalid input
        '404':
          description: Seat layout not found
        '401':
          description: Unauthorized
  /admin/seat-layouts/{event_id}/lock-seats:
    post:
      summary: Lock seats for an event (user-facing)
      tags:
        - Seat Layouts
      security:
        - bearerAuth: []
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                seatIds:
                  type: array
                  items:
                    type: string
              required:
                - seatIds
      responses:
        '200':
          description: Seats locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatLayout'
        '400':
          description: Invalid input
        '404':
          description: Seat layout not found
        '401':
          description: Unauthorized
  /admin/seat-layouts/{event_id}/unlock-seats:
    post:
      summary: Unlock seats for an event (user-facing)
      tags:
        - Seat Layouts
      security:
        - bearerAuth: []
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                seatIds:
                  type: array
                  items:
                    type: string
              required:
                - seatIds
      responses:
        '200':
          description: Seats unlocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatLayout'
        '400':
          description: Invalid input
        '404':
          description: Seat layout not found
        '401':
          description: Unauthorized
...
  /seat-layouts:
    post:
      summary: Create a public seat layout
      tags:
        - Public Seat Layouts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                event_id:
                  type: string
                layout_data:
                  type: array
                  items:
                    $ref: '#/components/schemas/SeatData'
              required:
                - event_id
                - layout_data
      responses:
        '201':
          description: Seat layout created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid input
  /seat-layouts/{event_id}:
    get:
      summary: Get public seat layout for an event
      tags:
        - Public Seat Layouts
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Seat layout details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/SeatLayout'
        '404':
          description: Seat layout not found
    put:
      summary: Update a public seat layout
      tags:
        - Public Seat Layouts
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                layout_data:
                  type: array
                  items:
                    $ref: '#/components/schemas/SeatData'
      responses:
        '200':
          description: Seat layout updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid input
        '404':
          description: Seat layout not found
    delete:
      summary: Delete a public seat layout
      tags:
        - Public Seat Layouts
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Seat layout deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Seat layout not found
  /seat-layouts/{event_id}/hold-seats:
    post:
      summary: Hold seats for a public event
      tags:
        - Public Seat Layouts
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                seatIds:
                  type: array
                  items:
                    type: string
              required:
                - seatIds
      responses:
        '200':
          description: Seats held successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid input
        '404':
          description: Seat layout not found
  /seat-layouts/{event_id}/release-seats:
    post:
      summary: Release held seats for a public event
      tags:
        - Public Seat Layouts
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                seatIds:
                  type: array
                  items:
                    type: string
              required:
                - seatIds
      responses:
        '200':
          description: Seats released successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid input
        '404':
          description: Seat layout not found
  /seat-layouts/{event_id}/publish:
    post:
      summary: Publish a public seat layout
      tags:
        - Public Seat Layouts
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Seat layout published successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Seat layout not found
  /seat-layouts/{event_id}/availability:
    get:
      summary: Get public seat availability
      tags:
        - Public Seat Layouts
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Seat availability details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/SeatLayout'
        '404':
          description: Seat layout not found
  /seat-layouts/{id}/book:
    post:
      summary: Book seats for a public event
      tags:
        - Public Seat Layouts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                seatIds:
                  type: array
                  items:
                    type: string
              required:
                - seatIds
      responses:
        '200':
          description: Seats booked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid input
        '404':
          description: Seat layout not found
  /seat-layouts/{event_id}/lock-seats:
    post:
      summary: Lock seats for a public event
      tags:
        - Public Seat Layouts
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                seatIds:
                  type: array
                  items:
                    type: string
              required:
                - seatIds
      responses:
        '200':
          description: Seats locked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid input
        '404':
          description: Seat layout not found
  /seat-layouts/{event_id}/unlock-seats:
    post:
      summary: Unlock seats for a public event
      tags:
        - Public Seat Layouts
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                seatIds:
                  type: array
                  items:
                    type: string
              required:
                - seatIds
      responses:
        '200':
          description: Seats unlocked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid input
        '404':
          description: Seat layout not found
...
security:
  - bearerAuth: []